--- Assets/CookieClickerProject/Program/Scripts/CookieClickerScene/Facility/Soldier.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Soldiers : MonoBehaviour
-{
-    int _IncreaseSoldiers = 1;
-    int _Dexresesgold = 100;
-    private void OnMouseDown()
-    {
-
-        Debug.Log("生産");
-        /*goldのマネージャー*/  -= _Dexresesgold;
-        /*兵のマネージャー*/ += _IncreaseSoldiers;
-
-    }
-}
