+++ Assets/FortressFableProject/Program/Scripts/FortressFableScene/Core/Facility/GoldMine.cs.meta
