+++ Assets/FortressFableProject/Program/Scripts/FortressFableScene/Core/Facility/Soldier.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Soldiers : MonoBehaviour
+{
+    int _IncreaseSoldiers = 1;
+    int _Dexresesgold = 100;
+
+    private void OnMouseDown()
+    {
+        Debug.Log("生産");
+        /*
+         * += _Dexresesgold;
+         * -= _IncreaseSoldiers;
+         */
+    }
+}
