+++ Assets/FortressFableProject/Program/Scripts/FortressFableScene/Core/Facility/Soldier.cs.meta
